+++ lib/ink.coffee
@@ -1,4 +1,6 @@
 {CompositeDisposable} = require 'atom'
+etch = require 'etch'
+
 PaneItem = require './util/pane-item'
 Loading = require './util/loading'
 progress = require './util/progress'
@@ -17,7 +19,8 @@ goto = require './gotodef'
 
 module.exports = Ink =
   activate: ->
-    mod.activate() for mod in [PaneItem, Result, Docs, Console, PlotPane, Workspace, breakpoints]
+    etch.setScheduler(atom.views)
+    mod.activate() for mod in [PaneItem, Result, Docs, Console, PlotPane, breakpoints]
 
   deactivate: ->
     pkg = atom.packages.getActivePackage 'ink'
