+++ styles/workspace.less
@@ -1,7 +1,7 @@
 @import "syntax-variables";
 @import "ui-variables";
 
-ink-workspace {
+.ink-workspace {
 
   -webkit-user-select: none;
   cursor: default;
